Allow specifying run order for plugins/transforms.
By default it maintains the previous behaviour of running all transforms
in the order they are declared, followed by all plugins in the order
they are declared.

This adds a new `pipeline` option, which is an array of plugin/transform
names. The plugins/transforms will be run according to the order
specified in the pipeline option. If a step is not specified, it will
not be run.

Example:

```javascript
browserify: {
  options: {
    plugin: [['tsify', {target: 'ES5', removeComments: true}]],
    transform: [['browserify-ngannotate', {ext: '.ts'}]],
    pipeline: ['tsify', 'browserify-ngannotate']
  },
  build: {
    'output.js': ['main.ts', 'module.ts']
  }
}
```